requirements.txt
suds-jurko
pymysql
flask-mail
flask
pip insta
pycurl